Version Packages (#7)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to master, this PR
will be updated.


# Releases
## @coralogix/opentelemetry@0.2.0

### Minor Changes

-   91a28cb: Added CoralogixTransactionSampler

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>